requirements.txt
discord.py>=2.3.0
openai>=1.63.0
python-dotenv>=1.0.0
aiohttp>=3.11.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiosqlite>=0.20.0
pyjwt>=2.9.0
python-multipart>=0.0.12
httpx>=0.28.0
google-generativeai>=0.8.0
